firebase쿼리 사용 및 리펙토링

+++ src/components/Nweet.js
@@ -13,7 +13,9 @@ const Nweet = (props) => {
     const ok = window.confirm("삭제하시겠습니까?");
     if (ok) {
       await dbService.doc(`nweets/${id}`).delete();
-      await storageService.refFromURL(attachmentURL).delete();
+      if(attachmentURL){
+        await storageService.refFromURL(attachmentURL).delete();
+      }
     }
   };
